Closures! Closures have been added now something like this works ``` fun make_counter(){     let i = 0;         fun count(){ 	        i = i + 1; 		println(i); 	} 	return count; }
let counter = make_counter();
counter(); //1
counter(); //2
counter(); //3
```

Function Values have also been changed to get this to work.
Value::Function now holds a u64 id, this id corresponds to an id in a
hashmap containing id and functions in the main interpretor struct.
When a function is called it updates its id to point at a new function
with the updated closure.

Other: Some comments for clairty were added around functions and call.
This is something im gonna work on more. This projects getting kinda big
and sometimes going back to older code gets a little tricky. Also
preparing to get better error reporting and move as many errors as
possible to be found in the Parser, this push has moved return
statements outside of a function to be a parser side error rather than
interpretor.