Resolving and binding, ```
let a = "global":
{
	fun showa(){
		println(a);
	}

	showa(); //global
	let a = "block"
	showa(); //global
	println(a);//block
}
```

This was probably the hardest commit thus far and took a lot of
refactoring old code to get it all to work. Its postponed my plans for
better error reporting and structs. Ill add some comments throughout the
code in the comming days to get the project better organized.

Bugs and such are expected ill fix those later :)